🚧 wip: create search page; add search controller

+++ lib/app/domain/entities/volume_info.dart
@@ -12,7 +12,7 @@ class VolumeInfo {
   final int? pageCount;
   final String? printType;
   final List<String>? categories;
-  final double? averageRating;
+  final int? averageRating;
   final int? ratingsCount;
   final String? maturityRating;
   final bool? allowAnonLogging;
